+++ lib/src/http_worker/http_worker.dart
@@ -1,4 +1,3 @@
-export 'package:unwired/src/http_worker/debug_http_worker.dart';
 export 'package:unwired/src/http_worker/stub_http_worker.dart'
-    if (dart.library.io) 'package:unwired/src/http_worker/native_http_worker.dart'
+    if (dart.library.io) 'package:unwired/src/http_worker/io_http_worker.dart'
     if (dart.library.html) 'package:unwired/src/http_worker/web_http_worker.dart';